2006_12dec24
git-svn-id: https://svn.code.sf.net/p/vitral/code/vitral/trunk@59 9e65a30e-a369-483d-98ba-aeb80c2fce51

+++ vitral/src/vsdk/toolkit/common/Vector3D.java
@@ -69,7 +69,7 @@ public class Vector3D extends Entity
     public final void normalize() {
         double t = x*x + y*y + z*z;
         if ( Math.abs(t) < VSDK.EPSILON ) return;
-        if (t != 0 && t != 1) t = (double) (1 / Math.sqrt(t));
+        if (t != 0 && t != 1) t = (1.0 / Math.sqrt(t));
         x *= t;
         y *= t;
         z *= t;
@@ -80,7 +80,7 @@ public class Vector3D extends Entity
      * @return double
      */
     public final double length() {
-        return (double)Math.sqrt(x*x + y*y + z*z);
+        return Math.sqrt(x*x + y*y + z*z);
     }
 
     public final Vector3D add(Vector3D b)
